refactor(core): collapse boxplot geometry family (#986)
Merge the 12-file boxplot geometry tree into geometry-boxplot.ts and
geometry-boxplot-body.ts. Delete orphaned geometry-smooth-line-write.ts
(zero src/ importers). Rewrite internals-reaching tests to assert on
rendered boxplot batches, including ggplot2 median fatten = 2.

Closes #986.
