--- packages/core/src/pipeline/geometry-boxplot-body-layout.ts
@@ -1,74 +1,0 @@
-/**
- * Boxplot body pixel layout: centers, hinge rects, whiskers, and median lines.
- */
-import type { BoxplotParams } from "@ggsvelte/spec";
-
-import type { LayerFrame, PipelineWarning } from "./types.js";
-import type { Frame } from "./geometry-shared.js";
-import {
-  DEFAULT_BOXPLOT_WIDTH,
-  MAX_BOXPLOT_PANEL_FRAC,
-  removedWarning,
-} from "./geometry-shared.js";
-import {
-  createBoxplotBodyBuffers,
-  pushKeptBoxplotRow,
-  pushRemovedBoxplotRow,
-} from "./geometry-boxplot-body-layout-collect.js";
-import { finalizeBoxplotBodyLayout } from "./geometry-boxplot-body-layout-finalize.js";
-import type { BoxplotBodyLayout } from "./geometry-boxplot-body-layout-types.js";
-import { layoutBoxplotBodyRow } from "./geometry-boxplot-body-row.js";
-
-export type { BoxplotBodyLayout } from "./geometry-boxplot-body-layout-types.js";
-
-const DEFAULT_BOX_LINEWIDTH = 1;
-
-export function layoutBoxplotBody(
-  frame: LayerFrame,
-  fx: Frame,
-  warnings: PipelineWarning[],
-): BoxplotBodyLayout | null {
-  const { binding, n } = frame;
-  const box = frame.box;
-  if (
-    box === null ||
-    frame.ymin === null ||
-    frame.ymax === null ||
-    fx.xScale.type !== "band" ||
-    fx.yScale.type === "band"
-  ) {
-    return null;
-  }
-  const params = (binding.layer.params ?? {}) as BoxplotParams;
-  const widthParam = params.width ?? DEFAULT_BOXPLOT_WIDTH;
-  let widthFrac = widthParam * fx.xScale.step;
-  // Few categories make band.step large; cap the default so boxes stay
-  // distribution-shaped. Authors who set `width` keep the uncapped fraction.
-  if (params.width === undefined) {
-    widthFrac = Math.min(widthFrac, MAX_BOXPLOT_PANEL_FRAC);
-  }
-  const linewidth =
-    typeof binding.linewidth.constant === "number"
-      ? binding.linewidth.constant
-      : (params.linewidth ?? DEFAULT_BOX_LINEWIDTH);
-  const alpha =
-    typeof binding.alpha.constant === "number" ? binding.alpha.constant : (params.alpha ?? 1);
-  const yScale = fx.yScale;
-
-  const buffers = createBoxplotBodyBuffers();
-  // box.lower/middle/upper and frame.ymin/ymax are stat_boxplot's aggregate
-  // output over already-transformed y (scale-space); normalizeTransformed
-  // skips the forward so they are never transformed twice.
-  const yPx = (v: number) => fx.innerHeight - yScale.normalizeTransformed(v) * fx.innerHeight;
-
-  for (let row = 0; row < n; row++) {
-    const geom = layoutBoxplotBodyRow({ frame, fx, row, widthFrac, yPx });
-    if (geom === null) {
-      pushRemovedBoxplotRow(buffers);
-      continue;
-    }
-    pushKeptBoxplotRow(buffers, geom, row);
-  }
-  removedWarning(buffers.removed, binding.index, warnings);
-  return finalizeBoxplotBodyLayout(buffers, linewidth, alpha, params);
-}
